Render list using autosizer with react-window
