+++ package.json
@@ -28,7 +28,9 @@
     "react-select": "^4.3.0",
     "react-split-pane": "^0.1.92",
     "react-use": "^17.2.1",
+    "react-virtualized-auto-sizer": "^1.0.5",
     "react-window": "^1.8.6",
+    "react-window-infinite-loader": "^1.0.7",
     "styled-components": "^5.2.1",
     "typescript": "^4.1.5",
     "use-places-autocomplete": "^1.9.1",
