+++ src/components/mobile/ListPage.js
@@ -1,25 +1,14 @@
-import { ChevronRight, Star } from '@styled-icons/boxicons-solid'
 import { useContext, useEffect, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import styled from 'styled-components/macro'
 
 import { getLocations } from '../../utils/api'
 import MapContext from '../map/MapContext'
-import { theme } from '../ui/GlobalStyle'
-import ListEntry from '../ui/ListEntry'
-
-const convertMetersToMiles = (meters) => {
-  let miles = (meters * 0.000621371192).toString()
-  miles = miles.slice(0, miles.indexOf('.') + 3)
-  return miles
-}
+import List from '../ui/List'
 
 const ListPageContainer = styled.div`
   margin-top: 85px;
-
-  &:not(:last-child) {
-    border-bottom: 1px solid ${({ theme }) => theme.secondaryBackground};
-  }
+  height: 100%;
 `
 
 const ListPage = () => {
@@ -54,16 +43,7 @@ const ListPage = () => {
 
   return (
     <ListPageContainer>
-      {locations.map((location) => (
-        <ListEntry
-          key={location.id}
-          leftIcons={<Star size="16" />}
-          rightIcons={<ChevronRight size="16" color={theme.blue} />}
-          primaryText={location.type_names[0]}
-          secondaryText={`${convertMetersToMiles(location.distance)} miles`}
-          onClick={() => handleListEntryClick(location.id)}
-        />
-      ))}
+      <List locations={locations} handleListEntryClick={handleListEntryClick} />
     </ListPageContainer>
   )
 }
